+++ desc/compute/_basis_vectors.py
@@ -45,11 +45,34 @@ def _b(params, transforms, profiles, data, **kwargs):
     profiles=[],
     coordinates="rtz",
     data=["e_theta/sqrt(g)", "e_zeta"],
+    aliases=["grad(rho)"],
 )
 def _e_sup_rho(params, transforms, profiles, data, **kwargs):
     # At the magnetic axis, this function returns the multivalued map whose
     # image is the set { 𝐞^ρ | ρ=0 }.
     data["e^rho"] = cross(data["e_theta/sqrt(g)"], data["e_zeta"])
+    return data
+
+
+@register_compute_fun(
+    name="e^rho*sqrt(g)",
+    label="\\mathbf{e}^{\\rho} \\sqrt{g}",
+    units="m^{2}",
+    units_long="square meters",
+    description="Contravariant radial basis vector weighted by 3-D volume Jacobian",
+    dim=3,
+    params=[],
+    transforms={},
+    profiles=[],
+    coordinates="rtz",
+    data=["e_theta", "e_zeta"],
+    parameterization=[
+        "desc.equilibrium.equilibrium.Equilibrium",
+        "desc.geometry.surface.FourierRZToroidalSurface",
+    ],
+)
+def _e_sup_rho_sqrtg(params, transforms, profiles, data, **kwargs):
+    data["e^rho*sqrt(g)"] = cross(data["e_theta"], data["e_zeta"])
     return data
 
 
@@ -498,6 +521,7 @@ def _e_sup_rho_zz(params, transforms, profiles, data, **kwargs):
     profiles=[],
     coordinates="rtz",
     data=["e^theta*sqrt(g)", "sqrt(g)"],
+    aliases=["grad(theta)"],
 )
 def _e_sup_theta(params, transforms, profiles, data, **kwargs):
     data["e^theta"] = (data["e^theta*sqrt(g)"].T / data["sqrt(g)"]).T
@@ -937,6 +961,7 @@ def _e_sup_theta_zz(params, transforms, profiles, data, **kwargs):
     profiles=[],
     coordinates="rtz",
     data=["e_rho", "e_theta/sqrt(g)"],
+    aliases=["grad(zeta)"],
 )
 def _e_sup_zeta(params, transforms, profiles, data, **kwargs):
     # At the magnetic axis, this function returns the multivalued map whose
@@ -3328,22 +3353,21 @@ def _gradpsi(params, transforms, profiles, data, **kwargs):
     transforms={},
     profiles=[],
     coordinates="rtz",
-    data=["e_theta", "e_zeta", "|e_theta x e_zeta|"],
-    axis_limit_data=["e_theta_r", "|e_theta x e_zeta|_r"],
-    parameterization=[
-        "desc.equilibrium.equilibrium.Equilibrium",
-    ],
+    data=["e^rho*sqrt(g)", "|e_theta x e_zeta|"],
+    axis_limit_data=["e_theta_r", "e_zeta", "|e_theta x e_zeta|_r"],
+    parameterization=["desc.equilibrium.equilibrium.Equilibrium"],
 )
 def _n_rho(params, transforms, profiles, data, **kwargs):
     # Equal to 𝐞^ρ / ‖𝐞^ρ‖ but works correctly for surfaces as well that don't
     # have contravariant basis defined.
     data["n_rho"] = transforms["grid"].replace_at_axis(
-        safediv(cross(data["e_theta"], data["e_zeta"]).T, data["|e_theta x e_zeta|"]).T,
+        safediv(data["e^rho*sqrt(g)"], data["|e_theta x e_zeta|"][:, jnp.newaxis]),
         # At the magnetic axis, this function returns the multivalued map whose
         # image is the set { 𝐞^ρ / ‖𝐞^ρ‖ | ρ=0 }.
         lambda: safediv(
-            cross(data["e_theta_r"], data["e_zeta"]).T, data["|e_theta x e_zeta|_r"]
-        ).T,
+            cross(data["e_theta_r"], data["e_zeta"]),
+            data["|e_theta x e_zeta|_r"][:, jnp.newaxis],
+        ),
     )
     return data
 
@@ -3359,18 +3383,15 @@ def _n_rho(params, transforms, profiles, data, **kwargs):
     transforms={},
     profiles=[],
     coordinates="rtz",
-    data=["e_theta", "e_zeta", "|e_theta x e_zeta|"],
-    parameterization=[
-        "desc.geometry.surface.FourierRZToroidalSurface",
-    ],
+    data=["e^rho*sqrt(g)", "|e_theta x e_zeta|"],
+    parameterization=["desc.geometry.surface.FourierRZToroidalSurface"],
 )
 def _n_rho_FourierRZToroidalSurface(params, transforms, profiles, data, **kwargs):
     # Equal to 𝐞^ρ / ‖𝐞^ρ‖ but works correctly for surfaces as well that don't
     # have contravariant basis defined.
     data["n_rho"] = safediv(
-        cross(data["e_theta"], data["e_zeta"]).T, data["|e_theta x e_zeta|"]
-    ).T
-
+        data["e^rho*sqrt(g)"], data["|e_theta x e_zeta|"][:, jnp.newaxis]
+    )
     return data
 
 
@@ -3423,7 +3444,7 @@ def _n_rho_z(params, transforms, profiles, data, **kwargs):
     transforms={},
     profiles=[],
     coordinates="rtz",
-    data=["e_rho", "e_zeta", "|e_zeta x e_rho|"],
+    data=["e^theta*sqrt(g)", "|e_zeta x e_rho|"],
     parameterization=[
         "desc.equilibrium.equilibrium.Equilibrium",
     ],
@@ -3431,9 +3452,7 @@ def _n_rho_z(params, transforms, profiles, data, **kwargs):
 def _n_theta(params, transforms, profiles, data, **kwargs):
     # Equal to 𝐞^θ / ‖𝐞^θ‖ but works correctly for surfaces as well that don't
     # have contravariant basis defined.
-    data["n_theta"] = (
-        cross(data["e_zeta"], data["e_rho"]).T / data["|e_zeta x e_rho|"]
-    ).T
+    data["n_theta"] = data["e^theta*sqrt(g)"] / data["|e_zeta x e_rho|"][:, jnp.newaxis]
     return data
 
 
@@ -3459,12 +3478,16 @@ def _n_zeta(params, transforms, profiles, data, **kwargs):
     # Equal to 𝐞^ζ / ‖𝐞^ζ‖ but works correctly for surfaces as well that don't
     # have contravariant basis defined.
     data["n_zeta"] = transforms["grid"].replace_at_axis(
-        safediv(cross(data["e_rho"], data["e_theta"]).T, data["|e_rho x e_theta|"]).T,
+        safediv(
+            cross(data["e_rho"], data["e_theta"]),
+            data["|e_rho x e_theta|"][:, jnp.newaxis],
+        ),
         # At the magnetic axis, this function returns the multivalued map whose
         # image is the set { 𝐞^ζ / ‖𝐞^ζ‖ | ρ=0 }.
         lambda: safediv(
-            cross(data["e_rho"], data["e_theta_r"]).T, data["|e_rho x e_theta|_r"]
-        ).T,
+            cross(data["e_rho"], data["e_theta_r"]),
+            data["|e_rho x e_theta|_r"][:, jnp.newaxis],
+        ),
     )
     return data
 
@@ -3762,4 +3785,77 @@ def _e_alpha_rp(params, transforms, profiles, data, **kwargs):
 )
 def _e_alpha_rp_norm(params, transforms, profiles, data, **kwargs):
     data["|e_alpha|r,p|"] = jnp.linalg.norm(data["e_alpha|r,p"], axis=-1)
+    return data
+
+
+@register_compute_fun(
+    name="n_rho x grad(theta)",
+    label="\\Vert \\mathbf{e}^{\\rho} \\Vert^{-1} \\mathbf{e}^{\\rho} "
+    "\times \\mathbf{e}^{\\theta}",
+    units="m^{-1}",
+    units_long="inverse meters",
+    description="Flux surface gradient of poloidal angle.",
+    dim=3,
+    params=[],
+    transforms={},
+    profiles=[],
+    coordinates="rtz",
+    data=["e_zeta", "|e_theta x e_zeta|"],
+    parameterization=[
+        "desc.equilibrium.equilibrium.Equilibrium",
+        "desc.geometry.surface.FourierRZToroidalSurface",
+    ],
+)
+def _surface_gradient_theta(params, transforms, profiles, data, **kwargs):
+    data["n_rho x grad(theta)"] = (
+        data["e_zeta"] / data["|e_theta x e_zeta|"][:, jnp.newaxis]
+    )
+    return data
+
+
+@register_compute_fun(
+    name="n_rho x grad(zeta)",
+    label="\\Vert \\mathbf{e}^{\\rho} \\Vert^{-1} \\mathbf{e}^{\\rho} "
+    "\times \\mathbf{e}^{\\zeta}",
+    units="m^{-1}",
+    units_long="inverse meters",
+    description="Flux surface gradient of toroidal angle.",
+    dim=3,
+    params=[],
+    transforms={},
+    profiles=[],
+    coordinates="rtz",
+    data=["e_theta", "|e_theta x e_zeta|"],
+    axis_limit_data=["e_theta_r", "|e_theta x e_zeta|_r"],
+    parameterization=["desc.equilibrium.equilibrium.Equilibrium"],
+)
+def _surface_gradient_zeta(params, transforms, profiles, data, **kwargs):
+    data["n_rho x grad(zeta)"] = transforms["grid"].replace_at_axis(
+        safediv(-data["e_theta"], data["|e_theta x e_zeta|"][:, jnp.newaxis]),
+        lambda: -data["e_theta_r"] / data["|e_theta x e_zeta|_r"][:, jnp.newaxis],
+    )
+    return data
+
+
+@register_compute_fun(
+    name="n_rho x grad(zeta)",
+    label="\\Vert \\mathbf{e}^{\\rho} \\Vert^{-1} \\mathbf{e}^{\\rho} "
+    "\times \\mathbf{e}^{\\zeta}",
+    units="m^{-1}",
+    units_long="inverse meters",
+    description="Flux surface gradient of toroidal angle.",
+    dim=3,
+    params=[],
+    transforms={},
+    profiles=[],
+    coordinates="rtz",
+    data=["e_theta", "|e_theta x e_zeta|"],
+    parameterization=["desc.geometry.surface.FourierRZToroidalSurface"],
+)
+def _surface_gradient_zeta_FourierRZToroidalSurface(
+    params, transforms, profiles, data, **kwargs
+):
+    data["n_rho x grad(zeta)"] = safediv(
+        -data["e_theta"], data["|e_theta x e_zeta|"][:, jnp.newaxis]
+    )
     return data